Modified War game
Changed War so that when both players play cards of equal value, those cards
are put aside (on_table).  If, in the next round, the both players again play cards
of equal value, those cards are added to the on_table list.  This continues until a
player wins a round, in which case they also win all the cards on the table.